refactor: 반복되는  메소드 추출
