+++ src/main/java/subway/controller/SectionController.java
@@ -31,47 +31,24 @@ public class SectionController extends ManageController {
 
     @Override
     public boolean register() { // 노선, 역 이름, 순서 입력 받고 등록
-        String line = getLine();
+        String line = method.getLine();
         if(!sectionManager.isValid(line)){
             return false;
         }
 
-        String station = getStation();
-        int index = getIndex();
+        String station = method.getStation();
+        int index = method.getIndex();
 
         sectionManager.insertSection(station, index-1);
         response.printInfo("구간이 등록되었습니다.");
         return true;
     }
 
-    public String getStation() {
-        ask.orderWhere("역 이름");
-        try {
-            return br.readLine();
-        } catch (Exception e) {
-            subwayException.unexpected();
-        }
-        return null;
-    }
-
-    public int getIndex() { // util?
-        ask.orderWhere();
-        try {
-            String input = br.readLine();
-            if(subwayException.isBack(input) == true) return -1;
-            int index = Integer.parseInt(input);
-            return index;
-        } catch (Exception e) {
-            subwayException.checkCommand();
-        }
-        return -1;
-    }
-
     @Override
     public boolean delete() {
         try{
-            String line = getLine(DELETE);
-            String station = getStation(DELETE, "구간의 역");
+            String line = method.getLine(DELETE, "구간의 노선");
+            String station = method.getStation(sectionManager, DELETE, "구간의 역");
 
             if(sectionManager.isValid(line)){
                 sectionManager.delete(station);
@@ -82,37 +59,6 @@ public class SectionController extends ManageController {
             return false;
         }
         return true;
-    }
-    public String getLine() { // util 로 뺄까?
-        ask.orderWhere(LINE);
-        try {
-            String line = br.readLine();
-            return line;
-        } catch (Exception e) {
-            e.printStackTrace();
-            subwayException.unexpected();
-        }
-        return null;
-    }
-
-    public String getLine(String Target){
-        try{
-            ask.orderWhere(Target, "구간의 노선");
-            return br.readLine();
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    public String getStation(String function, String target){
-        try{
-            ask.orderWhere(function, target);
-            return br.readLine();
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return null;
     }
     @Override
     public boolean read() {return false;}
